requirements.txt
python-dotenv
typer
pydantic
pydantic-settings
rich
pywhatkit
